fix: real process scores as numeric(5,2) + delete report from ProcessDetail
- Migrate reports/task_history score columns from integer to numeric(5,2)
  so fractional scores (e.g. 99.98) are stored instead of truncated to 100
- Backfill reports.score from result_json.failure_pct/severity (same formula as
  calculate_score) and sync task_history; models updated to Numeric(5,2)
- Add admin-only trash button in ProcessDetail 'Reportes (n)' list to delete a
  report (DELETE /reports/{id} already admin-protected)

+++ backend/qdata/core/score.py
@@ -7,7 +7,7 @@ WEIGHTS = {
 }
 
 
-def calculate_score(results: list[RuleResult]) -> tuple[int, str]:
+def calculate_score(results: list[RuleResult]) -> tuple[float, str]:
     total_weight = 0
     total_penalty = 0
